+++ src/app/pages/solicitudes/listar-solicitudes/listar-solicitudes.component.html
@@ -1,1 +1,2 @@
-<angular2-smart-table [settings]="settings" [source]="data"></angular2-smart-table>
+<!-- <angular2-smart-table [settings]="settings" [source]="data"></angular2-smart-table> -->
+<app-crud-solicitudes></app-crud-solicitudes>
