+++ src/app/pages/solicitudes/solicitudes.module.ts
@@ -5,12 +5,14 @@ import { SolicitudesRoutingModule } from './solicitudes-routing.module';
 import { ListarSolicitudesComponent } from './listar-solicitudes/listar-solicitudes.component';
 import { SharedModule } from 'src/app/shared/shared-module.module';
 import { FormSolicitudComponent } from './form-solicitud/form-solicitud.component';
+import { CrudSolicitudesComponent } from './crud-solicitudes/crud-solicitudes.component';
 
 
 @NgModule({
   declarations: [
     ListarSolicitudesComponent,
-    FormSolicitudComponent
+    FormSolicitudComponent,
+    CrudSolicitudesComponent,
   ],
   imports: [
     CommonModule,
